Storing data implemented through user default and project done

+++ Todo-mvc/Views/ListView.swift
@@ -10,28 +10,36 @@ import SwiftUI
 struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     var body: some View {
-        List{
-            ForEach(listViewModel.items){ item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            }else {
+                List{
+                    ForEach(listViewModel.items){ item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
+                
+            }
+            
+        }.navigationTitle("Todo List 📝")
+            .toolbar {
+                ToolbarItem (placement: .navigationBarLeading){
+                    NavigationLink("Add", destination: AddView())
+                }
+                ToolbarItem (placement: .navigationBarTrailing){
+                    EditButton()
                 }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-        }
-        .listStyle(PlainListStyle())
-        .navigationTitle("Todo List 📝")
-        .toolbar {
-            ToolbarItem (placement: .navigationBarLeading){
-                NavigationLink("Add", destination: AddView())
-            }
-            ToolbarItem (placement: .navigationBarTrailing){
-                EditButton()
-            }
-        }
+        
     }
